review fixes

+++ database/migrations/2023_05_05_202911_delete_not_accepted_user_cards.php
@@ -11,13 +11,16 @@ return new class extends Migration
     public function up(): void
     {
         $userCardIds = DB::table('user_cards')
-            ->join('order_items','order_items.id','user_cards.order_item_id')
-            ->where('order_items.order_item_status_id',OrderItemStatus::NOT_ACCEPTED)
+            ->join('order_items','order_items.id', 'user_cards.order_item_id')
+            ->where('order_items.order_item_status_id', OrderItemStatus::NOT_ACCEPTED)
             ->select('user_cards.id as id')
             ->get()->pluck('id')->flatten();
 
-        DB::table('user_card_certificates')->whereIn('user_card_id',$userCardIds)->delete();
-        DB::table('user_cards')->whereIn('id',$userCardIds)->delete();
+        if (count($userCardIds) > 0)
+        {
+            DB::table('user_card_certificates')->whereIn('user_card_id', $userCardIds)->delete();
+            DB::table('user_cards')->whereIn('id', $userCardIds)->delete();
+        }
     }
 
     /**
